refactored tests
Historically, the test suite has been constructed by concatenating
the test suites of many different implementations of this functionality
in other languages.

Now is the time to refactor this and reorganize test cases.

* regrouped test cases in categories relative to the part being
validated
* assert the returned error (check that we indeed captured the invalid
part in the expected method)
* removed redundant test cases

* fixes #21
